+++ autopush/tests/test_integration.py
@@ -181,7 +181,7 @@ keyid="http://example.org/bob/keys/123;salt="XZwpw6o37R-6qoZjw6KwAw"\
                     "Content-Type": "application/octet-stream",
                     "Content-Encoding": "aesgcm-128",
                     "Encryption": self._crypto_key,
-                    "Encryption-Key":
+                    "Crypto-Key":
                         'keyid="a1"; key="JcqK-OLkJZlJ3sJJWstJCA"',
                 })
             body = data or ""
@@ -242,7 +242,7 @@ keyid="http://example.org/bob/keys/123;salt="XZwpw6o37R-6qoZjw6KwAw"\
                 "Content-Type": "application/octet-stream",
                 "Content-Encoding": "aesgcm-128",
                 "Encryption": self._crypto_key,
-                "Encryption-Key":
+                "Crypto-Key":
                     'keyid="a1"; key="JcqK-OLkJZlJ3sJJWstJCA"',
             })
         body = data or ""